requirements.txt
click==7.1.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pdf2image==1.12.1
Pillow==7.0.0
pytesseract==0.3.3
Werkzeug==1.0.0
